Add files via upload
Access http://sqlfiddle.com/ and copy the SQL code for the creation of all of your tables in the left pane labeled “Schema Panel” to test your code and confirm the tables are built correctly without errors. Click on “Build Schema”.  Once this is completed, do a screen capture of what is showing in the sqlfiddle environment including the schema success message and paste this in the word document with the project.